* remove constants.ts (#446)
